Add @nowarn annotations to prevent chisel itself from spewing warnigns about Chisel
