+++ src/test/scala/chiselTests/CompatibilityInteroperabilitySpec.scala
@@ -4,7 +4,10 @@ package chiselTests
 
 import scala.collection.immutable.ListMap
 
+import scala.annotation.nowarn
+
 // Keep Chisel._ separate from chisel3._ below
+@nowarn("msg=Chisel compatibility mode is deprecated")
 object CompatibilityComponents {
   import Chisel._
   import Chisel3Components._
@@ -74,6 +77,7 @@ object Chisel3Components {
   class Chisel3ModuleChiselRecordB extends Chisel3PassthroughModule(Flipped(new ChiselRecord))
 }
 
+@nowarn("msg=Chisel compatibility mode is deprecated")
 class CompatibilityInteroperabilitySpec extends ChiselFlatSpec {
 
   "Modules defined in the Chisel._" should "successfully bulk connect in chisel3._" in {
@@ -106,6 +110,7 @@ class CompatibilityInteroperabilitySpec extends ChiselFlatSpec {
       b.io <> a.io
       stop()
     })
+
     assertTesterPasses(new BasicTester {
       val a = Module(new Chisel3RecordModuleA)
       val b = Module(new Chisel3RecordModuleB)
