+++ src/test/scala/chiselTests/ModuleExplicitResetSpec.scala
@@ -4,6 +4,9 @@ package chiselTests
 
 import chisel3.stage.ChiselStage
 
+import scala.annotation.nowarn
+
+@nowarn("msg=Chisel compatibility mode is deprecated")
 class ModuleExplicitResetSpec extends ChiselFlatSpec {
 
   "A Module with an explicit reset in compatibility mode" should "elaborate" in {
